requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2024.1

# UI Framework (Bonus)
streamlit==1.30.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
